Use the default language as fallback. Do not fail on errors.
The default language is used as fallback when the requested language does not exist or the requested string is not translated.
Previous versions of the library were designed to fail on load/parse errors. This version tries to fallback to default if possible and stores the load status and errors in the internal state.